amfd: validate node for being nodegroup member [#1311]
Amfd is allowing node to be deleted even if node is being part of node group.
This creates database differences because that node is still part of node group,
but it doesn's exists.
When a nodegroup is created and node is added to it, Amfd checks whether
node exists or not. If it doesn't exist, then node addition fails.
This is done in the expectation that node should exists if it want to
be part of node group. This concept should go inline with deletion of node
as well and node deletion shouldn't be allowed until it is part of a nodegroup.

A valiadation has been added when node is being deleted for its membership
of a node group. If node is a part of membership, it can't be deleted. So,
to delete a node, first delete that node from node group.