osaf: During adminrestart of node directors, before re-instantiating kill them [#1326] The command $ amf-adm restart <DN name> is one way of administratively restarting an AMF component. As apart of this admin operation, AMF sends the component terminate callback to the PI components. It is up to the component to release all its resources and respond to AMF the status of its self-termination before exiting (typically) the process itself. After receiving the response from the component, AMF invokes the instantiation script of the component. During this time, it is possible that the previously running instance of the process (of this component) has not yet exited. This situation when there is already a running daemon/process and now a new instantiation is being attempted can cause the instantiation script to return failure. This patch creates temporary term_state_file from inside the component terminate callback of the node directors. In the instantiation scripts, a check is done to distinguish a a fresh instantiation versus an instantiation after a termination. If the term_state_file exists then it means, its an instantiation after termination. If so, just attempt to kill (using killproc) the process again before calling start_daemon.
Note: There has been mention of using start_daemon -f option which will
create another copy of the daemon if the previous daemon is still running.
Using this option may not be ideal for us as it can create any inconsistency
between the two daemons when using any resources and also, there is no proof
or documentation of start_daemon -f working successfully. This is even more
significant given that some distros are really slow in becoming LSB
compliant, particularly the start_daemon and the likes of it.