amfd: ignore unlocked node during unlock-in op on ng [#1389]
Create a nodegroup in locked-in state while all of its nodes are in unlocked
state. Now if unlock-in operation is performed on the nodegroup then, nodes
state will be updated as LOCKED.

AMF should ignore those nodes which are in unlocked state while performing
unlock-in operation on nodegroup. At the same time if unlock-in admin
operation is performed on a node in lockied-in state and its nodegroup in
lock-in state, then AMF should not instantiate any SU on this node.

Patch fixes both the problems by including checks on node admin states and
nodegroup admin state.