plm: don't restart EE in opensaf_reboot unless it is instantiated [#1415] Hard resetting the active controller causes the standby controller to attempt to restart the (now reset) active controller's EE. This times out after 10 s, which holds up role switch over.
The call to immlist to get the admin state of the EE in opensaf_reboot is
actually returning <Empty> in this case.

Since PLM restart will not restart an EE that is not in instantiated state,
add a check for this. In this case <Empty> is returned and the
opensaf_reboot_with_plm function is not called, which allows the switchover
to proceed immediately.