log: cannot delete value of logStreamFileFormat [#1490]
At start up, `logStreamFileFormat` attribute in configuration obj has <empty> value.
Means that, built-in file format will be used as default value for app stream.
If user sets an value to logStreamFileFormat, users can not delete the value of this attribute.

This patch makes the attribute value can be deleted.