amfd: support si-swap admin op for Nway model [#78]
Currently si-swap is allowed on a SI that is protected by a 2N SG.
As per spec, operation is valid for NWAY and NPM model also.
This patch supports si-swap admin op for Nway model with following restrictions:
1)Operation will be rejected if Equal distribution feature is enabled.
2)operation will be rejected if SIRankedSU is configured and AutoAdjust flag is
enabled. However it will be allowed if AutoAdjust is disabled.

Reason for rejecting the operation in above two cases:
In both of these cases AutoAdjust flag is eanbled and is used to partially
adjust the SG. At the same time, autoadjust feature is used without
using autoadjustprob timer. In the equal distribution cases both active
assignments and standby assignments are equally distributed among the
available SUs without waiting for any autoadjustprob timer expiry.
Thus if si-swap is supported in this case, SG will be adjusted to
same configuration.
Same explanation is valid for other case also.