amf: support RESTART admin op on non-restartable component [#334]
AMF currently support admin restart operation on a restartable
component(for which saAmfDisableRestart=false).

Patch supports it on a non-restartable component.
Patch for #315 contains changes in amfnd and amfd which are common
for all the three tickets viz. #315, #334 and #1455.

TODO: To take care of controller role change when operation in
progress.