imm: Support SA_IMM_ATTR_STRONG_DEFAULT flag [#1425]
This flag is only allowed to be set on attributes that have a default value.
Attributes with this flag will never be NULL.

Adding this flag to an existing attribute (schema change) will be allowed after checking for ccb interference and existing objects of the class.

When there's an attempt to set NULL value (or to delete all values with ATTR_VALUES_DELETE) to attributes with this flag, default value will be set.
The AttrMod will be changed to ATTR_VALUES_REPLACE (attrValuesNumber = 1, attrValues is default value) and sent to OI.

ImmAttrValue::operator= is used to assign value to ImmAttrValue.
This will also work with ImmAttrMultiValue (head value will be assigned).